VerticalBarChart: migrated enzyme testcases to react test cases
